requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
gunicorn==23.0.0
bcrypt==4.2.1
Pillow==11.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
